Refactor Authorization
