+++ src/main/java/org/hl7/davinci/authorization/AuthEndpoint.java
@@ -1,4 +1,4 @@
-package org.hl7.davinci.priorauth;
+package org.hl7.davinci.authorization;
 
 import java.math.BigInteger;
 import java.security.KeyFactory;
@@ -7,19 +7,15 @@ import java.security.SecureRandom;
 import java.security.interfaces.RSAPublicKey;
 import java.security.spec.InvalidKeySpecException;
 import java.security.spec.RSAPublicKeySpec;
-import java.text.SimpleDateFormat;
 import java.util.ArrayList;
 import java.util.Base64;
 import java.util.Collections;
-import java.util.Date;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.UUID;
 import java.util.logging.Level;
 import java.util.logging.Logger;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 
 import javax.naming.directory.InvalidAttributeValueException;
 import javax.servlet.http.HttpServletRequest;
@@ -30,6 +26,10 @@ import com.auth0.jwt.JWTVerifier;
 import com.auth0.jwt.algorithms.Algorithm;
 import com.auth0.jwt.exceptions.JWTVerificationException;
 
+import org.hl7.davinci.priorauth.App;
+import org.hl7.davinci.priorauth.Audit;
+import org.hl7.davinci.priorauth.Endpoint;
+import org.hl7.davinci.priorauth.PALogger;
 import org.hl7.davinci.priorauth.Audit.AuditEventOutcome;
 import org.hl7.davinci.priorauth.Audit.AuditEventType;
 import org.hl7.davinci.priorauth.Database.Table;
@@ -61,9 +61,6 @@ public class AuthEndpoint {
 
     static final String INVALID_REQUEST = "invalid_request";
     static final String INVALID_CLIENT = "invalid_client";
-    static final String ADMIN_TOKEN = "Y3YWq2l08kvFqy50fQJY";
-
-    static final int TOKEN_LIFE_MIN = 5;
 
     @PostMapping(value = "/register", consumes = { MediaType.APPLICATION_JSON_VALUE, "application/fhir+json" })
     public ResponseEntity<String> registerClient(HttpServletRequest request, HttpEntity<String> entity) {
@@ -133,7 +130,7 @@ public class AuthEndpoint {
         // Check scope
         String[] requestedScopes = scope.split(" ");
         for(String requestedScope : requestedScopes) {
-            if (!getSupportedScopes().contains(requestedScope)) {
+            if (!AuthUtils.getSupportedScopes().contains(requestedScope)) {
                 String message = "requested scope " + requestedScope + " is not supported";
                 logger.warning("AuthEndpoint::token:" + message);
                 new Audit(AuditEventType.REST, AuditEventAction.E, AuditEventOutcome.MAJOR_FAILURE, null, request, "POST /token" + requestQueryParams);
@@ -216,122 +213,11 @@ public class AuthEndpoint {
         Map<String, Object> response = new HashMap<>();
         response.put("access_token", authToken);
         response.put("token_type", "bearer");
-        response.put("expires_in", TOKEN_LIFE_MIN * 60);
+        response.put("expires_in", AuthUtils.TOKEN_LIFE_MIN * 60);
         response.put("scope", "system/*.read");
         return ResponseEntity.status(HttpStatus.OK).contentType(MediaType.APPLICATION_JSON)
                 .cacheControl(CacheControl.noStore()).header("Pragma", "no-cache")
                 .body(JSONObject.toJSONString(response));
-    }
-
-    /**
-     * Populate the Client table with default data
-     */
-    public static void PopulateClientTable() {
-        // Auto register DTR RI
-        String clientId = "797da153-9a36-4493-9910-10648a4deb03";
-        Organization organization = new Organization();
-        organization.setId(clientId);
-        organization.setName("MITRE");
-        ContactPoint telecom = new ContactPoint();
-        telecom.setValue("blangley@example.com");
-        organization.setTelecom(Collections.singletonList(telecom));
-
-        HashMap<String, Object> dataMap = new HashMap<>();
-        dataMap.put("id", clientId);
-        dataMap.put("jwks", "{\"keys\":[{\"ext\":true,\"kty\":\"RSA\",\"e\":\"AQAB\",\"kid\":\"3ab8b05b64d799e289e10a201786b38c\",\"key_ops\":[\"verify\"],\"alg\":\"RS384\",\"n\":\"52tcPrGJgzyGqjcUiHsbSk_PxQ7Uovz4saGxva3iyBoidsekonigJJ3LnFlHYb3vBa2NA-0GpX2E1KhNNcYWAWQFcu069zi0YZ_wWGn6PWZURuonUoKH4dGHggym3xxVUxuA8OPubGe5ji56eic4RPINg0z-TtPlS-H9dnDIVznRUTXf3fy2dqWMuTY4D2e4fXGII6OpFAsEyrOqIoR8pLWGu7AiQkothunopp9q_Gu2xqB6l8BNulsbiwsQMeRE-9SGfeFpyblHiizHDwSqeZ3iv49Ellk4yjmrf6wOaFA2IXRqL1cCLj86B6KIDrjdzOL4lOSiES-PclNpioG2rQ\"}]}");
-        dataMap.put("jwks_url", null);
-        dataMap.put("token", "Y3YWq2l08kvFqy50fQJY");
-        dataMap.put("organization", organization);
-        App.getDB().write(Table.CLIENT, dataMap);
-    }
-
-    /**
-     * Determine whether or not the access token in the Authorization header is authorized
-     * 
-     * @param request HttpServletRequest
-     * @return true if the access token is valid, false if invalid or no bearer token provided
-     */
-    public static boolean validateAccessToken(HttpServletRequest request) {
-        String accessToken = getAccessToken(request);
-        if (accessToken == null) return false;
-
-        if (accessToken.equals(ADMIN_TOKEN)) {
-            logger.fine("AuthEndpoint::validateAccessToken:Admin token used");
-            return true;
-        }
-
-        String timestamp = App.getDB().readString(Table.CLIENT, Collections.singletonMap("token", accessToken), "timestamp");
-        if (timestamp == null) return false;
-
-        // Validate token age
-        SimpleDateFormat formatter = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss");
-        Date issuedAt;
-		try {
-			issuedAt = formatter.parse(timestamp);
-		} catch (java.text.ParseException e) {
-            e.printStackTrace();
-            logger.log(Level.SEVERE, "AuthEndpoint::validateAccessToken:Error parsing date", e);
-            return false;
-        }
-        
-        Date now = new Date();
-        long tokenAgeMs = now.getTime() - issuedAt.getTime();
-
-        if (tokenAgeMs > TOKEN_LIFE_MIN * 60000) {
-            logger.severe("AuthEndpoint::validateAccessToken:Access Token expired. Max Age " + TOKEN_LIFE_MIN*60000 + " but token age is " + tokenAgeMs);
-            return false;
-        }
-
-        return true;
-    }
-
-    /**
-     * Helper method to get client id via the access token from the request
-     * 
-     * @param request HttpServletRequest
-     * @return clientId if it exists or a string explaining why it does not
-     */
-    public static String getClientId(HttpServletRequest request) {
-        String accessToken = getAccessToken(request);
-        if (accessToken == null) return "Unknown Client: No Access Token";
-
-        if (accessToken.equals(ADMIN_TOKEN)) return "Admin";
-
-        String clientId = App.getDB().readString(Table.CLIENT, Collections.singletonMap("token", accessToken), "id");
-        return clientId == null ? "Unknown Client: Invalid Access Token" : clientId;
-    }
-
-    /**
-     * Helper method to get a list of all the supported scopes
-     * 
-     * @return list of scopes supported
-     */
-    public static List<String> getSupportedScopes() {
-        ArrayList<String> scopes = new ArrayList<>();
-        scopes.add("system/*.read");
-        scopes.add("offline_access");
-        return scopes;
-    }
-
-    /**
-     * Helper method to read the access token from the Authorization: Bearer {token} header
-     * 
-     * @param request request to get the token from
-     * @return access token if it was included, otherwise null
-     */
-    private static String getAccessToken(HttpServletRequest request) {
-        String authHeader = request.getHeader("Authorization");
-        logger.log(Level.FINE, "AuthEndpoint::getAccessToken:Authorization header: " + authHeader);
-        if (authHeader != null) {
-            String regex = "Bearer (.*)";
-            Pattern pattern = Pattern.compile(regex);
-            Matcher accessTokenMatcher = pattern.matcher(authHeader);
-            if (accessTokenMatcher.find() && accessTokenMatcher.groupCount() == 1) {
-                return accessTokenMatcher.group(1);
-            }
-        }
-
-        return null;
     }
 
     /**
